fix:백엔드에 맞게 형식수정
