+++ src/models/Store.js
@@ -1,4 +1,5 @@
 import { INDUSTRY_LABELS } from '../const/industries';
+import { isValidPhoneNumber } from '../utils/formatUtils';
 
 /**
  * Store 모델 - 백엔드 StoreResponse와 일치
@@ -49,7 +50,7 @@ export class Store {
 
   // 유효성 검사
   isValid() {
-    return this.name && this.address && this.phoneNumber && this.industry;
+    return this.name && this.address && this.phoneNumber && this.industry && isValidPhoneNumber(this.phoneNumber);
   }
 
   // 에러 메시지 가져오기
@@ -58,6 +59,7 @@ export class Store {
     if (!this.name) errors.push('매장명을 입력해주세요.');
     if (!this.address) errors.push('주소를 입력해주세요.');
     if (!this.phoneNumber) errors.push('연락처를 입력해주세요.');
+    if (!isValidPhoneNumber(this.phoneNumber)) errors.push('올바른 연락처 형식을 입력해주세요. (예: 02-1234-5678)');
     if (!this.industry) errors.push('업종을 선택해주세요.');
     return errors;
   }
@@ -71,7 +73,7 @@ export class Store {
       businessNumber: this.businessNumber,
       latitude: this.latitude,
       longitude: this.longitude,
-      industry: this.industry
+      industry: INDUSTRY_LABELS[this.industry] || this.industry
     };
   }
 
@@ -83,7 +85,7 @@ export class Store {
       phoneNumber: this.phoneNumber,
       latitude: this.latitude,
       longitude: this.longitude,
-      industry: this.industry
+      industry: INDUSTRY_LABELS[this.industry] || this.industry
     };
   }
 }
